like in overview

+++ src/app/article/article.component.ts
@@ -1,3 +1,4 @@
+import { Router } from '@angular/router';
 import { Article } from '../models/Article';
 import { ArticleService } from '../services/article.service';
 import { Component, Input, OnInit } from '@angular/core';
@@ -16,7 +17,7 @@ export class ArticleComponent implements OnInit {
   _user: User;
   user: string;
   _like: Boolean= false;
-  constructor(private service: ArticleService, private userService: UserService) {
+  constructor(private service: ArticleService, private userService: UserService, private router: Router) {
     if(JSON.parse(localStorage.getItem('currentUser')))
     {
       this.user = JSON.parse(localStorage.getItem('currentUser')).username;
@@ -47,29 +48,35 @@ export class ArticleComponent implements OnInit {
   }
 
   addLike(){
-    if(this._like == false)
-    {
-      var model = {
-        username: this._user.username,
+    if(this._user.username != null){
+      if(this._like == false)
+      {
+        var model = {
+          username: this._user.username,
+        }
+        this.service.addLike(this.article._id, model).subscribe(data =>{
+          //this._article.likes.push(data.user);
+          console.log(data);
+        });
+        this._like = true;
+        this.article.likes.push(this._user.username);
       }
-      this.service.addLike(this.article._id, model).subscribe(data =>{
-        //this._article.likes.push(data.user);
-        console.log(data);
-      });
-      this._like = true;
-      this.article.likes.push(this._user.username);
+      else{
+        var model = {
+          username: this._user.username,
+        }
+        this.service.removeLike(this.article._id, model).subscribe(data =>{
+          console.log(data);
+        });
+        this._like = false;
+        this.article.likes.pop();
+      }
+     
+      console.log(this.article.likes);
     }
     else{
-      var model = {
-        username: this._user.username,
-      }
-      this.service.removeLike(this.article._id, model).subscribe(data =>{
-        console.log(data);
-      });
-      this._like = false;
-      this.article.likes.pop();
+      this.router.navigate(['/login']);
     }
    
-    console.log(this.article.likes);
   }
 }
